Initial commit of 'tempus'
tempus is the main time tracking program. This has some basic
functionality working and can actually be used.

It currently has a built in timer. You can set a company name, project
and sub project. You can save and edit these entries. They are saved to
a tokycabinet table database under ~/.local/share/tempus/tempus.tdb

There is the start of the ability to set a description but this is
currently disabled as it's not fully functional yet.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>