tempus: Set the day to start at 0430
It is common in such things to have a new day not start at midnight but
at some point after it. Here I have it start at 0430.

This can be adjusted by setting the new_day_offset variable to the
number of seconds past midnight that you'd like the day to start. A
setting of 0 would simply mean midnight.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>