tempus: Tidy up the tokyocabinet storage
Due to generating a new primary key each time we saved an entry we would
end up with all the old entries as well as the new one. This happened to
work out OK, as when adding them to the GTree old ones would be replaced
with the new ones and we'd end up with just the latest for each tempus_id
showing.

But lets clean this up by using the tempus_id as the primary key and by
doing that we don't need to store it as a column in the database.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>