tempus: Make the list entries immutable
Make the entry box's in the entries list un-editable and un-focusable.
In other words they are purely for display purposes and using gtk
entries as opposed to text views is simpler for these simple text
displays.

You can of course still edit the current days entries via the edit
button.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>