tempus: Load entries in descending date order
When loading entries from the .tdb we would just load them in the order
they were written. This happened to work, though they loaded in from
oldest to newest and we want to display them from newest to oldest so we
had to re-order them when loading.

So lets load them in the order we want them, TDBQOSTRDESC. This also
*slightly* simplifies the code in load_tempi().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>