tempus: Tidy up the list display
Only display the date for each day once, so we now have for each day; a
GtkSeparator, that days date then the list of entries for that day.

We factor out code from load_tempi() into its own function that can also
then be used in cb_save(), this creates the "date header" (one per day).

We also rename is_editable() to is_today() as this is really what this
function is about, determining if the given date is today (taking into
account the start of day offset).

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>