tempus: Hook up description
Re-enable the description entry box and wire it up.

The description box is a GtkTextView and its contents are saved in the
.tdb under the description column. Descriptions are also saved as the
tooltip text for each list entries containing box. Thus when hovering
the mouse over any entry that has a description, this will show up as a
tooltip. These tooltips are also set at load time.

It remains to be seen how well that will work in the long term, but it
seems to be generally OK.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>