tempus: Enable auto completion of the company name
At start up when loading the data from the .tdb we add all the company
names to a temporary GTree, this handily ensures that we end up with a
sorted list of unique names.

We then add these to a liststore which is hooked up as the source for
the company name auto completion.

Then we remove the GTree as it's no longer needed.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>