tempus: Allow an adjusted elapsed_seconds when saving
When saving we would just take the value of elapsed_seconds. If you
wanted to adjust that you would have to make the adjustment then
start/stop the timer.

This commit allows you to make an adjustment and immediately save it.

We take the elapsed_seconds update code from cb_start_timer() and put it
into its own function that we then call from cb_start_timer() &
cb_save().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>