tempus: Don't show entries older than 90 days
When there is a lot of entries, which means a lot of widgets, there is
some performance issue in that tempus can be slightly sluggish in some
operations. This particularly shows up when trying to copy-paste the
times from the log entries, there can be log between initiating the
paste and the contents showing up. Also there can be lag in widget
focus.

A quick fix for this is to limit the amount of log entry widgets in use
so we limit the shown entries to the last 90 days. That brings the
responsiveness back.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>