tempus: Add a title bar marker when in record mode
When in record mode as well displaying "Rec" also display a circle in
front of the "Tempus" text. This circle is unicode character "U+23FA
BLACK CIRCLE FOR RECORD"

This should just help show when in record mode when the window is
minimised and you just have the task bar entry to look at and if you
have a lot of windows the "Rec" text may not be shown.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>