tempus: Fix a string truncation compiler warning
When compiling under GCC 8.x I got the following

    tempus.c: In function ‘set_tempi_store’:
    tempus.c:579:48: warning: ‘/tempus.tdb’ directive output may be truncated writing 11 bytes into a region of size between 1 and 4096 [-Wformat-truncation=]
      snprintf(tempi_store, sizeof(tempi_store), "%s/tempus.tdb", tempi_dir);
                                                    ^~~~~~~~~~~
    tempus.c:579:2: note: ‘snprintf’ output between 12 and 4107 bytes into a destination of size 4096
      snprintf(tempi_store, sizeof(tempi_store), "%s/tempus.tdb", tempi_dir);
      ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

tempi_store is PATH_MAX bytes large. tempi_dir is also PATH_MAX bytes
large.

So it's technically true. We could end up filling tempi_dir which would
fill up tempi_store before we got to write the filename component of the
path. However this is fairly unlikely. So the simple fix is to define
tempi_dir as PATH_MAX - the length of the subsequent filename component.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>